requirements.txt
alabaster==0.7.10
amqp==2.2.2
analytics-python==1.1.0
anyjson==0.3.3
appdirs==1.4.0
Babel==2.1.1
backports-abc==0.4
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.5.0.1
beautifulsoup4==4.4.1
billiard==3.5.0.3
celery==4.2.2
certifi==2015.11.20.1
cffi==1.9.1
chardet==3.0.4
click==6.7
contextlib2==0.5.5
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.1.4
decorator==4.0.11
defusedxml==0.5.0
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.11.13
django-constance==2.2.0
django-cors-headers==2.1.0
django-debug-toolbar==1.4
django-extensions==2.0.7
django-grappelli==2.11.1
django-heroku-redisify==0.2.1
django-mptt==0.8.7
django-nested-admin==3.0.21
django-nose==1.4.5
django-polymorphic==0.8.1
django-redis-cache==1.5.2
django-report-utils==0.3.15
django-rest-swagger==2.2.0
django-reversion==2.0.13
django-storages==1.1.8
django-toolbelt==0.0.1
django-uuidfield==0.5.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
djangorestframework-xml==1.3.0
docopt==0.4.0
docutils==0.14
drf-haystack==1.5.6
easy-thumbnails==2.3
ecdsa==0.13
elasticsearch==6.1.1
elasticsearch-dsl==6.1.0
elasticstack==0.4.0
enum34==1.1.6
factory-boy==2.8.1
Faker==0.7.15
firebase-admin==3.2.1
first==2.0.1
flower==0.8.3
funcsigs==1.0.2
future==0.16.0
futures==3.0.3
gunicorn==19.3.0
httplib2==0.9.2
humanize==0.5.1
idna==2.2
imagesize==0.7.1
inflect==0.2.5
ipaddress==1.0.18
ipdb==0.10.2
ipython==5.2.2
ipython-genutils==0.1.0
itypes==1.1.0
jdcal==1.2
Jinja2==2.9.6
Logentries==0.7
mandrill==1.0.57
MarkupSafe==1.0
mock==2.0.0
newrelic==2.96.0.80
nose==1.3.7
oauth2==1.9.0.post1
olefile==0.44
openapi-codec==1.3.2
openpyxl==2.2.1
opentok==2.5.0
packaging==16.8
pathlib2==2.2.1
paypalrestsdk==1.12.0
pbr==3.0.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.0.0
pip-tools==1.8.0
prompt-toolkit==1.0.13
psycopg2==2.7.4
ptyprocess==0.5.1
pyasn1==0.2.2
pycparser==2.17
Pygments==2.2.0
PyJWT==1.6.1
pyOpenSSL==16.2.0
pyparsing==2.1.10
python-dateutil==2.4.2
python-jose==1.3.2
python-monkey-business==1.0.0
python-slugify==1.2.6
pytz==2018.4
PyYAML==3.11
redis==2.10.5
requests==2.21.0
simplegeneric==0.8.1
simplejson==3.14.0
singledispatch==3.4.0.3
six==1.10.0
snowballstemmer==1.2.1
sqlparse==0.1.19
static3==0.6.1
stream-django==1.5.0
stream-python==3.0.1
tornado==4.3
traitlets==4.3.1
typing==3.6.2
Unidecode==0.4.20
uritemplate==3.0.0
urllib3==1.22
vine==1.1.4
wcwidth==0.1.7
Werkzeug==0.11.15
whitenoise==3.3.1
xlrd==1.0.0
xlwt==1.2.0
vine==1.1.4
kombu==4.3.0
billiard==3.5.0.3